feat(models): convert linguist and microphone to models (view indep.)
